refactor: Update resolutionValue import and remove unused utils files
- Changed the import path for resolutionValue to './dot-edit-content-form-resolutions'.
- Deleted unused utils.ts and utils.spec.ts files to clean up the codebase and improve maintainability.
- Minor adjustments in dot-edit-content-text-field component for better input handling.
