+++ core-web/libs/edit-content/src/lib/fields/dot-edit-content-text-field/dot-edit-content-text-field.component.html
@@ -1,3 +1,4 @@
+@let field = $field();
 <input
     [id]="field.variable"
     [formControlName]="field.variable"
